Split test_cli into two tests
